+++ quantum/encoder.h
@@ -15,6 +15,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+
+
+
+
 #pragma once
 
 #include "quantum.h"
@@ -23,9 +27,11 @@ void encoder_init(void);
 bool encoder_read(void);
 
 bool encoder_update_kb(uint8_t index, bool clockwise);
-bool encoder_update_user(uint8_t index, bool clockwise);
+bool encoder_update_user(uint16_t index, bool clockwise);
 
 #ifdef SPLIT_KEYBOARD
 void encoder_state_raw(uint8_t* slave_state);
 void encoder_update_raw(uint8_t* slave_state);
 #endif
+
+
